+++ code/src/cljs/sixsq/nuvla/ui/credentials/utils.cljs
@@ -5,15 +5,22 @@
             [sixsq.nuvla.ui.utils.time :as time]))
 
 
-(defn db->new-swarm-credential
+(defn db->new-coe-credential
   [db]
   (let [name        (get-in db [::spec/credential :name])
         description (get-in db [::spec/credential :description])
-        subtype     (get-in db [::spec/credential :subtype])
+        parent      (get-in db [::spec/credential :parent])
+
+        ;; subtype of the credential has to match the subtype for the infra-service
+        infra-subtype (->> (::spec/infrastructure-services-available db)
+                           (filter #(= (:id %) parent))
+                           first
+                           :subtype)
+        subtype (str "infrastructure-service-" infra-subtype)
+
         ca          (get-in db [::spec/credential :ca])
         cert        (get-in db [::spec/credential :cert])
         key         (get-in db [::spec/credential :key])
-        parent      (get-in db [::spec/credential :parent] [])
         acl         (get-in db [::spec/credential :acl])]
     (-> {}
         (assoc :name name)
@@ -93,13 +100,94 @@
         (assoc-in [:template :parent] infrastructure-services))))
 
 
+(defn db->new-exoscale-credential
+  [db]
+  (let [name        (get-in db [::spec/credential :name])
+        description (get-in db [::spec/credential :description])
+        subtype     (get-in db [::spec/credential :subtype])
+        secret      (get-in db [::spec/credential :exoscale-api-secret-key])
+        key         (get-in db [::spec/credential :exoscale-api-key])
+        acl         (get-in db [::spec/credential :acl])]
+    (-> {}
+        (assoc :name name)
+        (assoc :description description)
+        (assoc-in [:template :href] (str "credential-template/store-" subtype))
+        (assoc-in [:template :exoscale-api-key] key)
+        (assoc-in [:template :exoscale-api-secret-key] secret)
+        (cond-> acl (assoc-in [:template :acl] acl)))))
+
+
+(defn db->new-amazonec2-credential
+  [db]
+  (let [name        (get-in db [::spec/credential :name])
+        description (get-in db [::spec/credential :description])
+        subtype     (get-in db [::spec/credential :subtype])
+        key         (get-in db [::spec/credential :amazonec2-access-key])
+        secret      (get-in db [::spec/credential :amazonec2-secret-key])
+        acl         (get-in db [::spec/credential :acl])]
+    (-> {}
+        (assoc :name name)
+        (assoc :description description)
+        (assoc-in [:template :href] (str "credential-template/store-" subtype))
+        (assoc-in [:template :amazonec2-access-key] key)
+        (assoc-in [:template :amazonec2-secret-key] secret)
+        (cond-> acl (assoc-in [:template :acl] acl)))))
+
+
+(defn db->new-azure-credential
+  [db]
+  (let [name        (get-in db [::spec/credential :name])
+        description (get-in db [::spec/credential :description])
+        subtype     (get-in db [::spec/credential :subtype])
+        subsciption (get-in db [::spec/credential :azure-subscription-id])
+        id          (get-in db [::spec/credential :azure-client-id])
+        secret      (get-in db [::spec/credential :azure-client-secret])
+        acl         (get-in db [::spec/credential :acl])]
+    (-> {}
+        (assoc :name name)
+        (assoc :description description)
+        (assoc-in [:template :href] (str "credential-template/store-" subtype))
+        (assoc-in [:template :azure-subscription-id] subsciption)
+        (assoc-in [:template :azure-client-id] id)
+        (assoc-in [:template :azure-client-secret] secret)
+        (cond-> acl (assoc-in [:template :acl] acl)))))
+
+
+(defn db->new-google-credential
+  [db]
+  (let [name        (get-in db [::spec/credential :name])
+        description (get-in db [::spec/credential :description])
+        subtype     (get-in db [::spec/credential :subtype])
+        username    (get-in db [::spec/credential :google-username])
+        project     (get-in db [::spec/credential :google-project])
+        id          (get-in db [::spec/credential :client-id])
+        secret      (get-in db [::spec/credential :client-secret])
+        token       (get-in db [::spec/credential :refresh-token])
+        acl         (get-in db [::spec/credential :acl])]
+    (-> {}
+        (assoc :name name)
+        (assoc :description description)
+        (assoc-in [:template :href] (str "credential-template/store-" subtype))
+        (assoc-in [:template :google-username] username)
+        (assoc-in [:template :google-project] project)
+        (assoc-in [:template :client-id] id)
+        (assoc-in [:template :client-secret] secret)
+        (assoc-in [:template :refresh-token] token)
+        (cond-> acl (assoc-in [:template :acl] acl)))))
+
+
 (defn db->new-credential
   [db]
   (let [subtype (get-in db [::spec/credential :subtype])]
     (case subtype
-      "infrastructure-service-swarm" (db->new-swarm-credential db)
+      "infrastructure-service-swarm" (db->new-coe-credential db)
+      "infrastructure-service-kubernetes" (db->new-coe-credential db)
       "infrastructure-service-minio" (db->new-minio-credential db)
       "infrastructure-service-vpn" (db->new-vpn-credential db)
+      "infrastructure-service-exoscale" (db->new-exoscale-credential db)
+      "infrastructure-service-amazonec2" (db->new-amazonec2-credential db)
+      "infrastructure-service-azure" (db->new-azure-credential db)
+      "infrastructure-service-google" (db->new-google-credential db)
       "infrastructure-service-registry" (db->new-registry-credential db)
       "generate-ssh-key" (db->new-ssh-credential db))))