+++ code/src/cljs/sixsq/nuvla/ui/utils/resource_details.cljs
@@ -156,7 +156,7 @@
 
 
 (defn detail-header
-  [{:keys [id created updated name description properties acl parent subtype] :as data}]
+  [{:keys [id created updated name description properties acl parent subtype method state] :as data}]
   (when data
     [cc/metadata
      {:title       (or name id)
@@ -173,6 +173,8 @@
              created (conj (metadata-row "created" (time/time-value created)))
              updated (conj (metadata-row "updated" (time/time-value updated)))
              subtype (conj (metadata-row "subtype" subtype))
+             method (conj (metadata-row "method" method))
+             state (conj (metadata-row "state" state))
              parent (conj (metadata-row "parent" parent)))]))